song-vocab/pyproject.toml
[project]
name = "song-vocab"
version = "0.1.0"
description = "song vocabulary generator"
requires-python = ">=3.13.1"  # This is the correct way to specify Python version
dependencies = [
    "lyricsgenius",
    "langdetect",
    "python-dotenv>=1.0.1",
    "duckduckgo-search",
    "requests",
    "langchain",
    "langchain-community",
    "fastapi>=0.112.0",
    "uvicorn>=0.30.0"
]